refactor: extract preprocessing module + tighten docstrings
Three follow-ons to Goykhman's review remarks on PR #12:

* Extract anchoring tooling from highlevel.py into a new module
  numbox/utils/preprocessing.py (r3284175462). Sphinx "Cache-anchor
  mechanism" section moves from the highlevel subsection into a new
  numbox.utils.preprocessing subsection in docs/numbox.utils.rst.
  Background material (CPython LOAD_CONST, 3.14 LOAD_SMALL_INT, and
  pointers to numba's cache-key construction + source-stamp check)
  added alongside.

* Trim store_at docstring in lowlevel.py to drop the "Generalizes
  the int32-specific store pattern used by errno_set" sentence
  (r3284183320). That documented the function's history, not its
  current contract.

* Move _c.py module-level remarks about extern-symbol refs, @proxy
  caching, and reference sources to a new "Bindings module
  conventions" section in docs/numbox.core.bindings.rst (r3284248914).
  Those remarks apply to every bindings module (_math, _sqlite,
  _stdio, etc.), not just _c. _c.py keeps a one-line docstring.

403 tests pass on Py 3.12 + numba 0.65.1 locally.
