+++ docs/numbox.utils.rst
@@ -52,6 +52,14 @@ Then in a *different module* main.py define::
         assert isclose(data_2.derive_output(), 2.17)
 
 
+.. automodule:: numbox.utils.highlevel
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+numbox.utils.preprocessing
+--------------------------
+
 Cache-anchor mechanism
 ''''''''''''''''''''''
 
@@ -67,11 +75,12 @@ Numba's per-overload cache key
     (sig, codegen.magic_tuple(), hash(co_code), hash(closure_cells))
 
 It hashes ``co_code`` only, not ``co_consts``. Python's
-``LOAD_CONST`` opcode encodes an *index* into ``co_consts`` rather
-than the value itself, so two methods differing only in a numeric
-literal (``return self.x + 1`` vs ``return self.x + 1000``) produce
-identical ``co_code``. With a shared anchor file both would resolve
-to the same ``cache_subpath`` (numba's
+`LOAD_CONST <https://docs.python.org/3/library/dis.html#opcode-LOAD_CONST>`_
+opcode encodes an *index* into ``co_consts`` rather than the value
+itself, so two methods differing only in a numeric literal
+(``return self.x + 1`` vs ``return self.x + 1000``) produce identical
+``co_code``. With a shared anchor file both would resolve to the same
+``cache_subpath`` (numba's
 ``_CacheLocator.get_suitable_cache_subpath`` derives the cache subdir
 from a hash of ``co_filename``) and the second to compile would
 silently load the first's binary. Per-content anchors segregate the
@@ -85,13 +94,20 @@ narrow failure mode protected by content-addressing is constant-only
 edits (numeric or string literals, default arg values) where
 ``co_code`` is identical across versions.
 
-Python 3.14's ``LOAD_SMALL_INT`` opcode inlines small integers
-directly into ``co_code``, narrowing the failure mode on that version
-to constants outside the inline range. Earlier supported versions
-(3.10--3.13) collide on any constant edit.
+Python 3.14's
+`LOAD_SMALL_INT <https://docs.python.org/3.14/whatsnew/3.14.html>`_
+opcode inlines small integers directly into ``co_code``, narrowing
+the failure mode on that version to constants outside the inline
+range. Earlier supported versions (3.10--3.13) collide on any
+constant edit.
+
+See also ``numba.core.caching.Cache._index_key`` and
+``numba.core.caching._SourceFileBackedLocatorMixin.get_source_stamp``
+in numba's source for the cache key construction and source-stamp
+validity check.
 
 
-.. automodule:: numbox.utils.highlevel
+.. automodule:: numbox.utils.preprocessing
    :members:
    :show-inheritance:
    :undoc-members:
